google_cloud_functions/post_sns/src/requirements.txt
flask
google-api-python-client
google-auth
google-cloud-secret-manager
requests
requests_oauthlib
